requirements.txt
confluent-kafka==2.3.0
openmeteo-requests==1.2.0
requests==2.31.0
requests-cache==1.2.0
retry-requests==2.0.0
pandas==2.2.0
psycopg2-binary==2.9.9
boto3==1.34.0
python-dotenv==1.0.0
streamlit==1.32.0
